move other operations into OperationHandler
